# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: API for WASAPhoto, a webapp for sharing photos.
  version: 3.1.0
servers:
  - url: http://localhost:3000
    description: For development purposes
security:
  - UsernameAuth: [] 
paths:
...
  /users/{userId}/change-username:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    put:
      tags: ["change username"]
      summary: Change username
      description: | 
        If the username inserted is different from the old one the username
        is changed, otherwise it stays the same
      operationId: setMyUserName
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newUsername:
                  $ref: "#/components/schemas/Username"
              required:
                - newUsername
  
      responses:
        '200':
          description: Username changed successfully
          
        '204':
          description: Username is the similar to the former one

        '409':
          description: Username is already taken by other user
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{userId}/photo:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    get:
      parameters:
        - $ref: "#/components/parameters/Page"
      tags: ["photos"]
      summary: Get a list of user photos
      description: Retrieve photos in reverse chronological order
      operationId: getUserPhotos
      
      responses:
        '200':
          description: A list of photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photos"
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
      
      
    post:
      tags: ["my photos"]
      summary: Upload a photo
      description: | 
        Upload a photo for the authenticated user. Once submitted the server uploads the photo
        to a dedicated server/website and attaches a new id, author's username, date, time, 
        and initiates likesAmount and commentsAmount to 0.
        
      operationId: uploadPhoto
      requestBody:
        content:
          image/*:
            schema:
              $ref: "#/components/schemas/RawPhoto"
        required: true

      responses:
        '201':
          description: Photo uploaded successfully
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      
    get:
      tags: ["search users"]
      summary: Get user profile
      description: Retrieve details of a specific user's profile. Even yourself
      operationId: getUserProfile
      
      responses:
        '200':
          description: User profile details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        
        '404':
          description: User not found
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{photoId}:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    get:
      tags: ["stream"]
      summary: Get the photo itself
      description: Get the pecific photo itself.
      operationId: getPhoto
      responses:
        '200':
          description: Details of the photo
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/RawPhoto"
        
        '204':
          description: Photo not found
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /photos/{photoId}/details:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    get:
      tags: ["stream"]
      summary: Get details of a specific photo
      description: Retrieve details of a specific photo.
      operationId: getPhotoDetails
      responses:
        '200':
          description: Details of the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        
        '204':
          description: Photo not found
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{photoId}/comment:
    parameters:
      - $ref: "#/components/parameters/PhotoID"

    get:
      tags: ["comments"]
      summary: Get comments
      description: Get a list of all comments under the photo
      operationId: getComments
      
      responses:
        '200': 
          description: A list of comments
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                  
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  
    post:
      tags: ["comments"]
      summary: Add a comment to a photo
      description: Add a comment to a specific photo.
      operationId: commentPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
              
      responses:
        '201':
          description: Comment added successfully
          
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /photos/{photoId}/comment/{commentId}:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
      - $ref: "#/components/parameters/CommentID"
      
    delete:
      tags: ["comments"]
      summary: Remove a comment from a photo provided that it is your photo
      description: |
        Remove a comment from a specific photo provided that you are the author
        of the photo under which the comment lies.
      operationId: uncommentPhoto
      
      responses:
        '200':
          description: Comment removed successfully
        
        '404': 
          description: Comment not found
        
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '500': { $ref: "#/components/responses/InternalServerError" }
components:
  securitySchemes:
    UsernameAuth:
      type: http
      scheme: bearer
  parameters:
    PhotoID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: photoId
      in: path
      required: true
      description: Photo ID
    CommentID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: commentId
      in: path
      required: true
      description: Comment ID
    UserID:
      schema:
        $ref: "#/components/schemas/UniqueID"
      name: userId
      in: path
      required: true
      description: User ID
...
    Page:
      name: page
      in: query
      required: false
      description: |
        The page number for paginating through the data. Each page contains a set number of items of interest.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    Unauthorized:
      description: The access token is missing or invalid
    Forbidden:
      description: You do not have access rights to the operation
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  schemas:
    UniqueID:
      description: UUID for any object needed
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      readOnly: true
      example: "550e8400-e29b-41d4-a716-446655440000"
    Username:
      type: string
      example: Maria
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 32 
    RawPhoto:
      description: binary of an image
      type: object
      properties:
        photoBinary:
          type: string
          minLength: 0
          format: binary
          example: "010010101"
    Timestamp:
      description: Date and time representation
      type: string
      format: date-time
      example: "2023-01-01T12:00:00Z"
      minLength: 20
      maxLength: 20
      readOnly: true
    Amount:
      description: Representation of an amount of things of interest
      type: integer
      minimum: 0
      example: 42
      readOnly: true
    User:
      description: a user profile containg username, photos, followers and followings
      type: object
      properties:
        userID: 
          $ref: "#/components/schemas/UniqueID"
        username: 
          $ref: "#/components/schemas/Username"
        photos:
          $ref: "#/components/schemas/Amount"
        followers:
          $ref: "#/components/schemas/Amount"
        followings:
          $ref: "#/components/schemas/Amount"
...
    Photo:
      description: | 
        Once a stream of photos or a photo is requested they should be requested by a 
        host via GET method. 

      type: object
      properties:
        photoID: 
          $ref: "#/components/schemas/UniqueID"
        author: 
          $ref: "#/components/schemas/UniqueID"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        likesAmount:
          $ref: "#/components/schemas/Amount"
        comments:
          $ref: "#/components/schemas/Comments"
    Photos:
      description: |
        paginated list of photo ids from which you can retrieve 
        a particular photo using /photos/{photoID} endpoint
      type: array
      minItems: 1
      maxItems: 10
      items:
        $ref: "#/components/schemas/Photo/properties/photoID"
    Comment:
      description: Representation of a single comment
      type: object
      properties:
        commentID:
          $ref: "#/components/schemas/UniqueID"
        author: 
          $ref: "#/components/schemas/UniqueID"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        commentText:
          type: string
          example: "You look good, my man :)"
      required:
        - commentText  
    Comments:
      description: paginated list of comments
      type: array
      minItems: 1
      maxItems: 25
      items:
        $ref: "#/components/schemas/Comment"