# service/api/getUserPhotos.go
package api

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/Rsen337/WASA-project-WASAPhoto-/service/api/reqcontext"
	"github.com/julienschmidt/httprouter"
)

// returns ids of all the photos that belong to the athenticated user
func (rt *_router) getUserPhotos(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	w.Header().Set("Content-Type", "application/json")

	userId := ps.ByName("userId")

	// check if the user authorized and authenticated
	reqId := getToken(r.Header.Get("Authorization"))
	valid := rt.isAuthorized(reqId, reqId)
	if valid != 0 {
		ctx.Logger.Info("uploadPhoto: isAuthorized isn't happy")
		w.WriteHeader(valid)
		return
	}

	// checks if the requesting user is banned
	isBanned, err := rt.db.IsBanned(userId, reqId)
	if err != nil {
		ctx.Logger.WithError(err).Error("IsBanned returns error")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
	if isBanned {
		w.WriteHeader(http.StatusForbidden)
		ctx.Logger.Info("is banned")
		return
	}

	// Parse the page query parameter from the request URL
	pageStr := r.URL.Query().Get("page")
	page, err := strconv.Atoi(pageStr)
	if err != nil || page <= 0 {
		// means that the page query parameter wasn't provided, hence set it to default 1
		ctx.Logger.WithError(err).Error("page query not provided")
		page = 1
	}

	var photoIds []string

	photoIds, err = rt.db.GetUserPhotos(userId, page)
	if err != nil {
		ctx.Logger.WithError(err).Error("rt.db.getUserPhotos returned an error")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	// Convert photo IDs to JSON and write response
	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(photoIds); err != nil {
		ctx.Logger.WithError(err).Error("Error encoding JSON")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

}